fix bug

+++ app/controllers/BuildingController.php
@@ -71,10 +71,12 @@ class BuildingController extends Controller {
         if ($_SERVER['REQUEST_METHOD'] == 'POST')  {
             $data = array_map(fn($element) => trim($element), $_POST ); // assign post to data, and make sure that every value is trimmeds
             $data["error"] = [];
+
+
             if(!empty($_FILES["main_image"]["name"])) { // main image validation
                 $errorMsg = File::validateFiles($_FILES["main_image"], "validateImage");
                 if(empty($errorMsg) ) {
-                    unlink(UPLOADS . '/' . $data["main_image"]);
+                    if(!empty($data["main_image"])) unlink(UPLOADS . '/' . $data["main_image"]);
                     $data["main_image"] = File::handleValidatedFiles($_FILES["main_image"], UPLOADS);
                 } 
                 else $data["error"]["main_image"] =  $errorMsg;
@@ -83,25 +85,18 @@ class BuildingController extends Controller {
             if (!empty($_FILES["file"]["name"][0]))  { // auxiliary_images validation
                 $errorMsg = File::validateFiles($_FILES["file"], "validateImage");
                 if(empty($errorMsg) ) {
-                    $images = explode(" ", $data["auxiliary_images"]);
-                    foreach($images as $image) {
-                        unlink(UPLOADS . '/' . $image);
+                    if(!empty($data["auxiliary_images"]))  {
+                        $images = explode(" ", $data["auxiliary_images"]);
+                        foreach($images as $image) {
+                            unlink(UPLOADS . '/' . $image);
+                        }
                     }
                     $data["auxiliary_images"] = File::handleValidatedFiles($_FILES["file"], UPLOADS);
                 } 
                 else $data["error"]["auxiliary_images"] =  $errorMsg;
             }
-            if(!empty($_FILES["main_image"]["name"])) { // main image validation
-                $errorMsg = File::validateFiles($_FILES["main_image"], "validateImage");
-                if(empty($errorMsg) ) $data["main_image"] = File::handleValidatedFiles($_FILES["main_image"], UPLOADS);
-                else $data["error"]["main_image"] =  $errorMsg;
-            }
             
             if (empty($data["error"])) {
-                echo "<pre>";
-                print_r($data["error"]);
-                echo "</pre>";
-                exit();
                 $model = $this->model(MODELS_NAMESPACE . "BuildingModel");
                 if(call_user_func_array([$model, "update"], [$data])) {
                     MessageReporting::flash(MessagesName::Building, "building updated succfully");
